# [LOG4J2-928] Mostly wait-free MemoryMappedFileManager