… log4j-core/src/main/java/org/apache/logging/log4j/core/appender/OutputStreamManager.java
… log4j-core/src/main/java/org/apache/logging/log4j/core/appender/ThreadHints.java
… log4j-core/src/main/java/org/apache/logging/log4j/core/appender/ThreadHintsMH.java
… log4j-core/src/main/java/org/apache/logging/log4j/core/appender/UninterruptibleCountDownLatch.java
… log4j-core/src/main/java/org/apache/logging/log4j/core/appender/UnlockedAwaitableReentrantLock.java
… log4j-core/src/main/java/org/apache/logging/log4j/core/appender/UnsafeMappedByteBufferWriteTouch.java
… log4j-core/src/main/java/org/apache/logging/log4j/core/layout/ByteBufferDestination.java
… log4j-core/src/main/java/org/apache/logging/log4j/core/layout/LockableByteBufferDestination.java
… log4j-core/src/main/java/org/apache/logging/log4j/core/layout/LockingStringBuilderEncoder.java
… log4j-core/src/main/java/org/apache/logging/log4j/core/layout/StringBuilderEncoder.java
… log4j-core/src/main/java/org/apache/logging/log4j/core/layout/TextEncoderHelper.java
… log4j-core/src/test/java/org/apache/logging/log4j/core/appender/MemoryMappedFileManagerConcurrentWriteTest.java
… log4j-core/src/test/java/org/apache/logging/log4j/core/appender/MemoryMappedFileManagerTest.java
… log4j-core/src/test/java/org/apache/logging/log4j/core/async/perftest/IdleStrategy.java
… log4j-core/src/test/java/org/apache/logging/log4j/core/async/perftest/Pacer.java
… log4j-core/src/test/java/org/apache/logging/log4j/core/async/perftest/ResponseTimeTest.java
… log4j-core/src/test/java/org/apache/logging/log4j/core/async/perftest/RunLog4j2.java
… log4j-core/src/test/java/org/apache/logging/log4j/core/async/perftest/YieldIdleStrategy.java
… log4j-core/src/test/java/org/apache/logging/log4j/core/layout/PatternLayoutTest.java
… log4j-perf/pom.xml
… log4j-perf/src/main/java/org/apache/logging/log4j/perf/IdleStrategy.java
… log4j-perf/src/main/java/org/apache/logging/log4j/perf/NoOpIdleStrategy.java
… log4j-perf/src/main/java/org/apache/logging/log4j/perf/Pacer.java
… log4j-perf/src/main/java/org/apache/logging/log4j/perf/YieldIdleStrategy.java
… log4j-perf/src/main/java/org/apache/logging/log4j/perf/jmh/ByteArrayToByteBufferCopyBenchmark.java
… log4j-perf/src/main/java/org/apache/logging/log4j/perf/jmh/Log4j2AppenderComparisonBenchmark.java
… log4j-perf/src/main/java/org/apache/logging/log4j/perf/jmh/Log4j2AppenderLatencyBenchmark.java
… log4j-perf/src/main/java/org/apache/logging/log4j/perf/jmh/Log4j2AppenderThroughputBenchmark.java
… pom.xml