A pull request for https://issues.apache.org/jira/browse/LOG4J2-928.

This PR is not complete - it lacks tests and benchmarks (I'm going to add them in the next few weeks), but the core implementation is reviewable already.