@remkop 2017-07-01
Roman, one more thing: since this is a non-trivial patch, would you mind submitting an [ICLA](http://www.apache.org/licenses/#clas)?