@leventov 2017-07-01
@remkop 

> MemoryMappedFileManager::switchRegion and ::closeAndSwitchRegionWhileLocked should not throw Error: generally a logging problem should not bring down the application. See also the {{ignoreExceptions}} [configuration|https://logging.apache.org/log4j/2.x/manual/appenders.html#MemoryMappedFileAppender] attribute that allows applications to choose to receive logging exceptions.

I follow the approach of java.util.concurrent internals here, where `throw new Error(...)` is used in the face of unrecoverable and totally unexpected errors. the alternative is `AssertionError`, or `assert` statements.