@leventov 2017-07-01
@remkop note that this PR is primarily about _latency_, not throughput at all. I will do some testing, but if the throughput will degrade (that I expect in a single-thread case), and latency with improve, I consider this a success.